codegen: Refactor AbstractTemplateRenderingPass
AbstractTemplateRenderingPass now inherits from the new AbstractMultiTemplateRenderingPass, which allows rendering multiple files from a single template.
